エラー修復(floatからintへの変換エラー)

+++ SuperNewRoles/Roles/Role/RoleClass.cs
@@ -1231,8 +1231,8 @@ namespace SuperNewRoles.Roles
                     LevelingerPlayer = new();
                     ThisXP = 0;
                     IsCreateMadmate = false;
-                    OneKillXP = CustomOptions.LevelingerOneKillXP.GetFloat();
-                    UpLevelXp = CustomOptions.LevelingerUpLevelXP.GetFloat();
+                    OneKillXP = CustomOptions.LevelingerOneKillXP.GetInt();
+                    UpLevelXp = CustomOptions.LevelingerUpLevelXP.GetInt();
                     GetPowerData = new();
                     for (int i = 0; i < 5; i++)
                     {
@@ -1245,7 +1245,7 @@ namespace SuperNewRoles.Roles
                         GetPowerData.Add(GetLevelPowerType(getdata));
                     }
                     IsUseOKRevive = CustomOptions.LevelingerReviveXP.GetBool();
-                    ReviveUseXP = CustomOptions.LevelingerUseXPRevive.GetFloat();
+                    ReviveUseXP = CustomOptions.LevelingerUseXPRevive.GetInt();
                 }
                 catch { }
             }
@@ -1369,7 +1369,7 @@ namespace SuperNewRoles.Roles
             public static void ClearAndReload()
             {
                 MadMayorPlayer = new();
-                AddVote = CustomOptions.MadMayorVoteCount.GetFloat();
+                AddVote = CustomOptions.MadMayorVoteCount.GetInt();
                 IsImpostorCheck = CustomOptions.MadMayorIsCheckImpostor.GetBool();
                 IsUseVent = CustomOptions.MadMayorIsUseVent.GetBool();
                 IsImpostorLight = CustomOptions.MadMayorIsImpostorLight.GetBool();
@@ -2096,7 +2096,7 @@ namespace SuperNewRoles.Roles
                     Short = PlayerControl.GameOptions.NumShortTasks;
                 }
                 JackalCheckTask = (int)(AllTask * (int.Parse(CustomOptions.MayorFriendsCheckJackalTask.GetString().Replace("%", "")) / 100f));
-                AddVote = CustomOptions.MayorFriendsVoteCount.GetFloat();
+                AddVote = CustomOptions.MayorFriendsVoteCount.GetInt();
             }
         }
         public static class VentMaker
@@ -2136,7 +2136,7 @@ namespace SuperNewRoles.Roles
             public static void ClearAndReload()
             {
                 GhostMechanicPlayer = new();
-                LimitCount = CustomOptions.GhostMechanicRepairLimit.GetFloat();
+                LimitCount = CustomOptions.GhostMechanicRepairLimit.GetInt();
             }
         }
         public static class EvilHacker
